Re-arranged address info component props
