--- services/explorer/src/api/mocks/events.ts
@@ -1,11 +1,0 @@
-export const getMockedEvent = () => {
-  const event = [
-    {
-      name: 'Transfer',
-      from: '0x00e193a15486909eba3fb36c815cb8a331180cc97a27ffb69b8122de02e5ea18',
-      to: '0x00e193a15486909eba3fb36c815cb8a331180cc97a27ffb69b8122de02e5ea18',
-      value: '100000',
-    },
-  ];
-  return event;
-};
