--- services/explorer/src/api/mocks/index.ts
@@ -1,4 +1,0 @@
-export * from './transactions';
-export * from './blocks';
-export * from './nodes';
-export * from './events';
