--- services/explorer/src/api/mocks/nodes.ts
@@ -1,133 +1,0 @@
-import { ColumnData, NodesTableData } from '../../models';
-
-export const getMockedNodesColsAndRows = () => {
-  const cols: ColumnData[] = [
-    { path: 'rank', name: 'Rank' },
-    { path: 'address', name: 'Node Address' },
-    { path: 'blocksProduced', name: '# blocks produced' },
-  ];
-
-  const rows: NodesTableData[] = [
-    {
-      rank: 1,
-      address: '0xc26f6b31a5f8452201af8db5cc25cf4340df8b09',
-      pbftCount: 3213,
-    },
-    {
-      rank: 2,
-      address: '0xc26f6b31a5f8452201af8db5cc25cf4340df8b09',
-      pbftCount: 3213,
-    },
-    {
-      rank: 3,
-      address: '0xc26f6b31a5f8452201af8db5cc25cf4340df8b09',
-      pbftCount: 3213,
-    },
-    {
-      rank: 4,
-      address: '0xc26f6b31a5f8452201af8db5cc25cf4340df8b09',
-      pbftCount: 3213,
-    },
-    {
-      rank: 5,
-      address: '0xc26f6b31a5f8452201af8db5cc25cf4340df8b09',
-      pbftCount: 3213,
-    },
-    {
-      rank: 6,
-      address: '0xc26f6b31a5f8452201af8db5cc25cf4340df8b09',
-      pbftCount: 3213,
-    },
-    {
-      rank: 7,
-      address: '0xc26f6b31a5f8452201af8db5cc25cf4340df8b09',
-      pbftCount: 3213,
-    },
-    {
-      rank: 8,
-      address: '0xc26f6b31a5f8452201af8db5cc25cf4340df8b09',
-      pbftCount: 3213,
-    },
-    {
-      rank: 9,
-      address: '0xc26f6b31a5f8452201af8db5cc25cf4340df8b09',
-      pbftCount: 3213,
-    },
-    {
-      rank: 10,
-      address: '0xc26f6b31a5f8452201af8db5cc25cf4340df8b09',
-      pbftCount: 3213,
-    },
-    {
-      rank: 11,
-      address: '0xc26f6b31a5f8452201af8db5cc25cf4340df8b09',
-      pbftCount: 3213,
-    },
-    {
-      rank: 12,
-      address: '0xc26f6b31a5f8452201af8db5cc25cf4340df8b09',
-      pbftCount: 3213,
-    },
-    {
-      rank: 13,
-      address: '0xc26f6b31a5f8452201af8db5cc25cf4340df8b09',
-      pbftCount: 3213,
-    },
-    {
-      rank: 14,
-      address: '0xc26f6b31a5f8452201af8db5cc25cf4340df8b09',
-      pbftCount: 3213,
-    },
-    {
-      rank: 15,
-      address: '0xc26f6b31a5f8452201af8db5cc25cf4340df8b09',
-      pbftCount: 3213,
-    },
-    {
-      rank: 16,
-      address: '0xc26f6b31a5f8452201af8db5cc25cf4340df8b09',
-      pbftCount: 3213,
-    },
-    {
-      rank: 17,
-      address: '0xc26f6b31a5f8452201af8db5cc25cf4340df8b09',
-      pbftCount: 3213,
-    },
-    {
-      rank: 18,
-      address: '0xc26f6b31a5f8452201af8db5cc25cf4340df8b09',
-      pbftCount: 3213,
-    },
-    {
-      rank: 19,
-      address: '0xc26f6b31a5f8452201af8db5cc25cf4340df8b09',
-      pbftCount: 3213,
-    },
-    {
-      rank: 20,
-      address: '0xc26f6b31a5f8452201af8db5cc25cf4340df8b09',
-      pbftCount: 3213,
-    },
-    {
-      rank: 21,
-      address: '0xc26f6b31a5f8452201af8db5cc25cf4340df8b09',
-      pbftCount: 3213,
-    },
-    {
-      rank: 22,
-      address: '0xc26f6b31a5f8452201af8db5cc25cf4340df8b09',
-      pbftCount: 3213,
-    },
-    {
-      rank: 23,
-      address: '0xc26f6b31a5f8452201af8db5cc25cf4340df8b09',
-      pbftCount: 3213,
-    },
-    {
-      rank: 24,
-      address: '0xc26f6b31a5f8452201af8db5cc25cf4340df8b09',
-      pbftCount: 3213,
-    },
-  ];
-  return { cols, rows };
-};
