+++ services/explorer/src/pages/AddressInfo/AddressInfo.tsx
@@ -1,6 +1,7 @@
+import { Box, Paper } from '@mui/material';
 import React from 'react';
 import { useParams } from 'react-router-dom';
-import { AddressInfo, PageTitle } from '../../components';
+import { AddressDetails, AddressTables, PageTitle } from '../../components';
 import { useAddressInfoEffects } from './AddressInfo.effects';
 import AddressLoadingSkeleton from './AddressLoadingSkeleton';
 
@@ -31,7 +32,6 @@ const AddressInfoPage = (): JSX.Element => {
     setTabsStep,
     isFetchingAddressStats,
     isLoadingAddressStats,
-    isLoadingTables,
   } = useAddressInfoEffects(account);
 
   return (
@@ -43,32 +43,43 @@ const AddressInfoPage = (): JSX.Element => {
       {showLoadingSkeleton ? (
         <AddressLoadingSkeleton />
       ) : (
-        <AddressInfo
-          details={addressInfoDetails}
-          transactions={transactions}
-          dagBlocks={dagBlocks}
-          pbftBlocks={pbftBlocks}
-          totalPbftCount={totalPbftCount}
-          rowsPbftPerPage={rowsPbftPerPage}
-          pbftPage={pbftPage}
-          handlePbftChangePage={handlePbftChangePage}
-          handlePbftChangeRowsPerPage={handlePbftChangeRowsPerPage}
-          totalDagCount={totalDagCount}
-          rowsDagPerPage={rowsDagPerPage}
-          dagPage={dagPage}
-          handleDagChangePage={handleDagChangePage}
-          handleDagChangeRowsPerPage={handleDagChangeRowsPerPage}
-          totalTxCount={totalTxCount}
-          rowsTxPerPage={rowsTxPerPage}
-          txPage={txPage}
-          handleTxChangePage={handleTxChangePage}
-          handleTxChangeRowsPerPage={handleTxChangeRowsPerPage}
-          tabsStep={tabsStep}
-          setTabsStep={setTabsStep}
-          isFetchingAddressStats={isFetchingAddressStats}
-          isLoadingAddressStats={isLoadingAddressStats}
-          isLoadingTables={isLoadingTables}
-        />
+        <Paper elevation={1}>
+          <Box
+            display='flex'
+            flexDirection='column'
+            alignItems='left'
+            margin='2rem 2rem 2rem'
+            gap='1.5rem'
+          >
+            <AddressDetails
+              details={addressInfoDetails}
+              isFetchingAddressStats={isFetchingAddressStats}
+              isLoadingAddressStats={isLoadingAddressStats}
+            />
+            <AddressTables
+              transactions={transactions}
+              dagBlocks={dagBlocks}
+              pbftBlocks={pbftBlocks}
+              totalPbftCount={totalPbftCount}
+              rowsPbftPerPage={rowsPbftPerPage}
+              pbftPage={pbftPage}
+              handlePbftChangePage={handlePbftChangePage}
+              handlePbftChangeRowsPerPage={handlePbftChangeRowsPerPage}
+              totalDagCount={totalDagCount}
+              rowsDagPerPage={rowsDagPerPage}
+              dagPage={dagPage}
+              handleDagChangePage={handleDagChangePage}
+              handleDagChangeRowsPerPage={handleDagChangeRowsPerPage}
+              totalTxCount={totalTxCount}
+              rowsTxPerPage={rowsTxPerPage}
+              txPage={txPage}
+              handleTxChangePage={handleTxChangePage}
+              handleTxChangeRowsPerPage={handleTxChangeRowsPerPage}
+              tabsStep={tabsStep}
+              setTabsStep={setTabsStep}
+            />
+          </Box>
+        </Paper>
       )}
     </>
   );
